Tweak handcart qol (#1627)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- Note: PRs including balance changes authored by anyone other than
maintainers and official devs will not be considered. -->

## About The Pull Request

This adds some small QoL changes for handcarts, which I think make the
process of mining a little more fun.
Handcarts will now attempt to filter for a few commonly gathered types
of objects for their `attack_hand` proc. This is mostly meant to address
a common problem of having to load the dross rock alongside the valuable
ores you unearth when mining, needlessly cluttering the handcart with
trash rocks.

If you click the cart with nothing in your hand, the cart will first
seek out any ores or gems on the turf beneath the user. It will gather
them and only them, if they exist.

Next, it moves on to wood -- specifically, logs. It does the same
filtering thing there.

Then, it continues the default logic of grabbing everything it can on
the turf.

I can easily add some more common use cases if I've missed a big one,
but the primary issue was mixing ores with trash rocks.

## Why It's Good For The Game

This makes the mining experience less tedious by removing a common pain
point that I've encountered.

## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [x] You tested this on a local server.
- [x] This code did not runtime during testing.
- [x] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->