Adv mage nerf + mage apprentice buff (#1346)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- Note: PRs including balance changes authored by anyone other than
maintainers and official devs will not be considered. -->

## About The Pull Request

+ removes the third big quarz from the aprentice pouch, because it
doesnt even spawn because too big

Adv mage
+ Reduces the starting spell points of avd mage/sorcceres from 10 to 8
+ Removes the "learn spell" spell from sorcceres she had for some reason
+ Gives Adv mage/sorcceres a poorly filled component pouch, wich 2 mana
bloom, 2 big quarz, 2 small quarz

Mage apprentice
+ Adds Prestidigitation to mage apprentice
+ Adds filled component pouch to mage apprentice
+ Adds a chemise shirt to female mage apprentice
+ Adds chalk to the mage apprentice satchel

Mapping
+ Changes Vanderlin.dmm lower mage guild doors to use tower lock id

## Why It's Good For The Game

Sligth rebalance of adv mages and mage apprentice.

Adv mages are a bit too strong currently. They start of ready to go and
not needing the guild at all. Even though joining the guild should be
giving them a rather big advantage. They are still better then
aprentices in a actual abilitys, just worse in supplys, being they arent
living in a mage guild but traveling.

Most materials are pretty easy to gather, so i removed the easiest from
the pouch for now. Its been reduced to the more troubelsome components
only. Quarz and mana bloom. But they get one more quarz, because that is
currently the main bottleneck. It should give them a better chance at
being solo, without being able to just ignore the mage guild.

Adv mage players bosted a lot about how strong they are from the get go,
and how cool it is to have roundstart meteor storm and more.
Bad idea to brag about that sorta stuff i tell you.
8 rounstart points + the ones from the book is MORE then enough.

Along with that mage apprentice get a bit of an easier live.
They are meant to be the role used to learn the magic system, yet they
have it the worst.

So, they gain a full pouch to freely experiment with rituals and chalk.
Along with that they gain the spell Prestidigitation at roundstart.
This way mage apprentice will be ready to go, roundstart.
Being able to immediatly experiment with the magic system.

Also, the lower mage guild on Vanderlin still needed court mages key,
wich already was changed in rosewood.
The whole point of the predrawn summoning rune is to intice unfamilar
players to use it, doesnt make sense to make them unable to even see it.

Discord for balance discussion: NHC


![grafik](https://github.com/user-attachments/assets/3c264369-ac64-405c-8491-aa2415ba368a)


## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [x] You tested this on a local server.
- [x] This code did not runtime during testing.
- [x] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->