Resprites dwarves. (#1719)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- Note: PRs including balance changes authored by anyone other than
maintainers and official devs will not be considered. -->

## About The Pull Request

![DwarfIngameClothed](https://github.com/user-attachments/assets/d038d612-16fb-4e1a-bee6-5ab818b7989d)

![DwarfResprite](https://github.com/user-attachments/assets/e6c53a53-cc4d-4617-9059-6337d71cae37)

![DwarfIngame](https://github.com/user-attachments/assets/3e7cfee7-c31c-4a1c-813d-3964322af2e6)

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review or prevent the PR from being merged!
-->

## Why It's Good For The Game
Dwarves are the only species who still have their old models. It looks
strikingly ugly and out of place when looking at all characters in
comparison. This PR gives them some much-needed TLC and brings their
body sprites to the same style as the others.
<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [x] You tested this on a local server.
- [x] This code did not runtime during testing.
- [x] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->

Co-authored-by: AberrantQuerist <crystalseeker04@gmail.com>