Drow Knight and Squire sexism update (#1708)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- Note: PRs including balance changes authored by anyone other than
maintainers and official devs will not be considered. -->

## About The Pull Request
The Zizo specific ZIZO KNIGHT can now only be female
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review or prevent the PR from being merged!
-->

## Why It's Good For The Game
Zizo sexist tbh
Also Gale (assistant headmin) told me to make this.
Will ask Aberra for permission as well rq.
Update: [Lore lead
approved.](https://discord.com/channels/748354466335686736/1309303420661792801/1366571564539183125)
<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [x] You tested this on a local server.
- [x] This code did not runtime during testing.
- [x] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->