Eora's Day PR - Adds Boquets, Flower Crowns and Roses! - Courtesy of Lutowski from Azure Peak (#1456)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- Note: PRs including balance changes authored by anyone other than
maintainers and official devs will not be considered. -->

## About The Pull Request

Adds pretty flowers, bouqets and flower crowns, all craftable

To craft any of them, you must add the flower of your choice to a fibre,
and then for a bouquet, finish it off with another flower and a cloth.

Sprites and general (edited) code courtesy of Lutowski from Azure Peak
Original PR can be found at
https://github.com/GeneralPantsuIsBadAtCoding/Azure-Peak/pull/1410


## Why It's Good For The Game

Pretty flowers, whats not to love


![dreamseeker_tdiO6K08Ie](https://github.com/user-attachments/assets/97c4c591-4f78-4b2a-ba3d-63dea5fdf94b)

![dreammaker_PSjKZIfO6d](https://github.com/user-attachments/assets/5d926bc6-6f62-492f-a7bd-dc297e17920b)


## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [x] You tested this on a local server.
- [x] This code did not runtime during testing.
- [ ] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->

---------

Co-authored-by: grungussuss <96586172+grungussuss@users.noreply.github.com>