Map switch again but without cbt (#1569)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- Note: PRs including balance changes authored by anyone other than
maintainers and official devs will not be considered. -->

## About The Pull Request
This one https://github.com/Vanderlin-Tales-Of-Wine/Vanderlin/pull/1257
but using SPACEMAN_DMM which means map icons only work in StrongDMM but
that's okay because why wouldn't you be using StrongDMM? Borbop said
this is fine.

Don't merge until
https://github.com/Vanderlin-Tales-Of-Wine/Vanderlin/pull/1510 is merged
because that reverts 1257

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review or prevent the PR from being merged!
-->

## Why It's Good For The Game
Check the other one

<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [x] You tested this on a local server.
- [x] This code did not runtime during testing.
- [x] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->