Vanderlin Train Station Rework (#1700)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- Note: PRs including balance changes authored by anyone other than
maintainers and official devs will not be considered. -->

## About The Pull Request

Finishes the train station up and makes the train as long as a real-life
train should be. All carriages function as far travel points.

## Why It's Good For The Game

Looks nice, Vanderlin now has an actual train

There are a few loot spawns, they consist of basic alchemical
ingredients, and a few food spawns, aswell as a small amount of coal and
coal dust.

![StrongDMM_knB7ygqEt0](https://github.com/user-attachments/assets/3274cbbb-c4a1-43c4-9579-c64cf27f4fad)

![StrongDMM_JA8QOBIEy1](https://github.com/user-attachments/assets/4e6f3acf-a583-4d64-81e7-c66b88072a65)
Images of both the train platform and the new aboveground section


## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [x] You tested this on a local server.
- [x] This code did not runtime during testing.
- [x] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->