Daftmarsh (#1639)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- Note: PRs including balance changes authored by anyone other than
maintainers and official devs will not be considered. -->

## About The Pull Request
"All the kings said it was DAFT to build a castle in the swamp, but I
built it all the same! Just to show 'em! ...It sank into the swamp- SO!
I built a second one! That sank into the swamp. So I built a third one.
That burned down, fell over, and then sank into the swamp. BUT THE
FOURTH ONE STAYED UP!!"
-Monty Python and the Holy Grail


![image](https://github.com/user-attachments/assets/91839522-96ea-4e64-bc61-7db1c2589ae8)

Daftmarsh is a lower population based map which invokes more of a
'small-town' feel compared to the city. It is (as of writing) still work
in progress, and the outlands section is woefully incomplete.

Daftmarsh is located just to the North of Vanderlin, a trade town
constructed on stilts atop the ruins of a half-sunken castle. Whilst
they may call themselves kings (for now,) the Count/Countess of
Daftmarsh are adaptable if nothing else. It may be a shithole, but it's
their shithole. And a comfy one, at that!
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review or prevent the PR from being merged!
-->

## Why It's Good For The Game
Vanderlin needs a lower population map with a more compact design for
the evenings/European hours.
<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [x] You tested this on a local server.
- [x] This code did not runtime during testing.
- [x] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->

---------

Co-authored-by: AberrantQuerist <crystalseeker04@gmail.com>
Co-authored-by: grungussuss <96586172+grungussuss@users.noreply.github.com>