The Joy of Tea. (potentially) PART 1 (#1706)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- Note: PRs including balance changes authored by anyone other than
maintainers and official devs will not be considered. -->

## About The Pull Request
Expands cooking by an addition of different teas with properties to help
fatigued fellows in their everydae life.
Not so cheap or fast method to get some early game simple heals, mana
restoration and some more.

## Why It's Good For The Game
1. Expands cooking options.
2. Its no secret herbal teas where used as remedies for different
ailments. Concidering how underfeatured medicine is, having
nonoverpowered medicinal teas will cover some of the issues with healing
people, mainly giving wounded some time to be brought to propper healing
means, or at least, enough vigor to crawl to safety if one is lucky.
3. Despite being pretty much very minor, DIY potions, they will come in
handy for a plethora of roles, mainly those of cooking, surviving in the
wilds, pilgrim alchemists who cant afford a cauldron (they will be able
to sell atleast something this basic to combat classes).
4. More food preservation, a mediocre niche of drinks (sits between
plain water and wines)
5. Balancewise - all teas are 500%-300% weaker than potions with same
effects. This, arguably only not true to swampweed brew, which has no
counterpart and is essentially a much thicker concotion with higher
concentration of swampweed essence

![image](https://github.com/user-attachments/assets/ee26b2a1-6187-48a6-82d2-72896dbe566b)



## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [x] You tested this on a local server.
- [x] This code did not runtime during testing.
- [x] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->