Allows all species(except Half-Orcs, Hollow-kin and Kobolds) to be guards. (#1765)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- Note: PRs including balance changes authored by anyone other than
maintainers and official devs will not be considered. -->

## About The Pull Request
Title, Allows Guards to be any races except Half-Orcs, Hollow-kin and
Kobold
Along with that, Allows AGE_IMMORTAL on Forest Guard, because Aasimar
couldn't be forest guard, for whatever reason??
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review or prevent the PR from being merged!
-->

## Why It's Good For The Game
I don't get why they can't be guards, I could understand Man at arms,
but City Guards? I feel like it is strange, it'll also give more guard
players so it's a plus-plus.. was also talked during the town-hall and
could be a good idea..
<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [x] You tested this on a local server.
- [x] This code did not runtime during testing.
- [x] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->